moved Search files and DELETED searchBar expander
the expander allows the toolbar to show a search button which,
when clicked, prompts the user to search.
This behavior mimicks the "search" button from Youtube

It is no more used, and reference apps don't use this kind of behavior,
but this long commit message is to be able to find it back if needed.

this intermediate commit will not compile.