reset password page when user opens mail link
when clicking on the link, a prompt with Polyvox Logo appears,
and when the user selects it, he goes to the app to change his password.

For now, the server answers me that the id is invalid when sending
request, I checked it, I think it's a bug, or I have a too old mail.