File Viewers for images, PDF, Web links and others
massive commit that contains 4 features in one.
by calling 1 function, FileLoader.openFile
(name: String, url: String, type: String = "", act: Activity)
the type is deduced from the URL if it is left empty

it opens a new fragment with:

an ImageView
a PDF reader (for phones with API >= 21)
a Webview to see web links
a prompt to open an unknown file with an app on the phone.
(exemple: an .xls file shows that it can be opened with Powerpoint)

a footer contains 3 icons allowing:
sharing all the documents, and also directly share the pictures.
downloading the PDFs and images.
settings documents as "favorite". not implemented yet, as not decided in
API.

this commit also introduces a FileProvider, required to share the documents.