chat can be used in landscape mode, can be hidden.
just like Twitch chat on mobile.
the animation to show/hide the chat in landscape is not very pretty,
but making it good would require changing all the layout and code on
this part...
there is still the option to remove the animation and show/hide
directly, which neither is pleasant to watch, but at least it's instant.
it's in a all caps comment in the code

added animation library unused for the moment.