🎉version 0.5: all functionalities implemented!🎉
not all functionalities are connected to the API,
but dummy behavior have been implemented while waiting the creations of
the last API routes.

this commit introduces another useless lib that takes to much size:
AboutLibraries.
It shows lib licences, and an "About this app" section.
I added ourselves in credits!