added global search results view
when searching, you now get a page with the 3 first results of rooms and
3 first user results.
The user can still go on other tabs to get all the rooms and users
results.

This commit took a particularly long time to do, as "a view that
contains just 2 fragments of variable size" is easier sid than done.

Contains an ugly hack where I calculate the size in pixels of my
fragments to manually resize the layout.

but now everything works!