:camera: special commit: camera streaming attempt.
this commit will be pushed on a special branch/

It contains all the experiments with Android's MediaRecorder and Camera2
API in order to implement the Speaker functionality for Polyvox.

Unfortunately, the MediaRecorder API is flawed, and video recording
with sound didn't worked on 2 of our testing machines (Honor 7 and 8).
Recording with WebM don't have any sound, and recording in other formats
make the app crash.

Moreover, streaming video content to the server seems very complicated
as the MediaRecorder API is not conceived for streaming, splitting the
video in chuncks would be an hard task.

Not pictured in this commit, but I spent 1 day trying to integrate the
libstreaming (https://github.com/fyhertz/libstreaming) library to the project.
The import failed, as a strange Gradle error said it couldn't find a
plugin "needed for libstreaming", even if it was referenced nowhere in
the code.

moreover, after torough check of the libstreaming code,
it seemed it would have the same issues than my previous implementation,
and it couldn't work with Websockets without complex changes.

I spent a lot of time on this feature, and I don't exclude coming back
at it, but it is useless for the app right now, sadly.