removed unused UI elements, updated UI, fixed chat
removed "join" button on a follower that is not in a room.
removed dummy room when not having network connection
added number of actors in room description.
removed references and buttons leading to private chat
added a "/broadcast" prefix to message to comply with chat change
changed version number to 0.66