added localized date formatter method
getPrettyDate transforms a Date object to a localized string like
"yesterday 10:00" , "monday 14:30" or "March 1 2017".

it looks small when reading it again, but it took me some time to do it.
the unit test is longer than the method it tests...
it isn't even really necessary... but at least I did it!