premium user!
user can become subscribe for a premium service.

premium users have a small medal label on their picture in
their profile page.
removed ads for premium users.

added A Dialog layout to prompt the user to subscribe
this prompt will show ocasionally on the home page if the user is not
premium.

updated API to take in account this feature.