refactored and reimplemented chat
completely reworked the CustomChat class for the private chat.

Transformed class to Singleton.
Created methods to get and send private messages,
and get list of the people with which you sent messages.

Reworked Room and private chat code to use this new implementation.
threw away all the sweet code designed for complex features that will
never see the day in the private chat (file and picture upload...)

it was not wort it, but did it anyway.